utils/pathology_processor.py
import cv2
import numpy as np
from PIL import Image
import os
from typing import Dict, Any

class PathologyImageProcessor:
    """
    Real pathology processor using OpenCV and NumPy.
    Analyzes cellularity and staining intensity in histopathology slides.
    """
    
    def analyze_tissue(self, image_path: str) -> Dict[str, Any]:
        """Performs real color-space and morphology analysis on tissue slides"""
        if not os.path.exists(image_path):
            return {}
            
        try:
            # Load image using OpenCV
            img = cv2.imread(image_path)
            if img is None:
                return {}
                
            # Convert to HSV for better color segmentation (H&E stain analysis)
            hsv = cv2.cvtColor(img, cv2.COLOR_BGR2HSV)
            
            # Hematoxylin (Purple/Blue - Nuclei)
            lower_nuclei = np.array([120, 50, 50])
            upper_nuclei = np.array([160, 255, 255])
            mask_nuclei = cv2.inRange(hsv, lower_nuclei, upper_nuclei)
            
            # Eosin (Pink - Cytoplasm/Stroma)
            lower_cyto = np.array([160, 50, 50])
            upper_cyto = np.array([180, 255, 255])
            mask_cyto = cv2.inRange(hsv, lower_cyto, upper_cyto)
            
            nuclei_density = np.sum(mask_nuclei > 0) / mask_nuclei.size
            cyto_density = np.sum(mask_cyto > 0) / mask_cyto.size
            
            # Morphological analysis (feature extraction)
            kernel = np.ones((5,5), np.uint8)
            opening = cv2.morphologyEx(mask_nuclei, cv2.MORPH_OPEN, kernel)
            contours, _ = cv2.findContours(opening, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
            
            return {
                "nuclei_density": float(nuclei_density),
                "cellular_atypia_index": float(cyto_density / (nuclei_density + 1e-6)),
                "estimated_cell_count": len(contours),
                "avg_nuclei_size": float(np.mean([cv2.contourArea(c) for c in contours])) if contours else 0.0
            }
        except Exception as e:
            return {"error": str(e)}
