requirements.txt
# Core Dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
sqlalchemy>=2.0.23

# Streamlit UI
streamlit>=1.29.0

# Medical Imaging
pydicom>=2.4.4
opencv-python>=4.8.1.78
numpy>=1.26.0
Pillow>=10.1.0

# Knowledge & RAG
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Utilities
python-dateutil>=2.8.2
pypdf2>=3.0.1
reportlab>=4.0.0

# Development
pytest>=7.4.3
black>=23.11.0
